requirements.txt
asgiref==3.3.1
beautifulsoup4==4.9.3
cssutils==2.2.0
Django==3.1.7
django-ckeditor==6.0.0
django-cleanup==5.1.0
django-etc==1.3.0
django-hitcount==1.3.3
django-js-asset==1.2.2
future==0.18.2
Pillow
pynliner==0.8.0
pytz==2021.1
soupsieve==2.2.1
sqlparse==0.4.1